separate components in main folders
